% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confband.R
\name{confband}
\alias{confband}
\title{Function to plot the confidence bands}
\usage{
confband(xV, yVmin, yVmax)
}
\arguments{
\item{xV}{the values for the x-axis.}

\item{yVmin}{the minimum values for the y-axis.}

\item{yVmax}{the maximum values for the y-axis.}
}
\value{
a polygon.
}
\description{
Function to plot the confidence bands
}
