% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fdaSP.internal}
\alias{fdaSP.internal}
\alias{.lm_dof_LASSO_1lambda}
\alias{.lm_dof_OVGLASSO_1lambda}
\alias{.lm_cov_dof_OVGLASSO_1lambda}
\alias{.lm_adaptive_dof_OVGLASSO_1lambda}
\alias{.lm_cov_adaptive_dof_OVGLASSO_1lamba}
\alias{.f2s_dof_1lambda}
\alias{.f2s_cov_dof_1lambda}
\alias{.f2s_smo_dof_1lambda}
\alias{.f2s_cov_smo_dof_1lambda}
\alias{.f2s_adaptive_dof_1lambda}
\alias{.f2s_cov_adaptive_dof_1lambda}
\alias{.f2s_smo_adaptive_dof_1lambda}
\alias{.f2s_adaptive_cov_smo_dof_1lambda}
\alias{.f2s_bic}
\alias{.f2s_cov_bic}
\alias{.f2s_ebic}
\alias{.f2s_cov_ebic}
\alias{.admm_glasso}
\alias{.admm_glasso_cov}
\alias{.admm_glasso_fast}
\alias{.admm_glasso_cov_fast}
\alias{.admm_spglasso}
\alias{.admm_spglasso_cov}
\alias{.admm_spglasso_fast}
\alias{.admm_spglasso_cov_fast}
\alias{.admm_lasso}
\alias{.admm_lasso_cov}
\alias{.admm_adalasso}
\alias{.admm_adalasso_cov}
\alias{.admm_lasso_fast}
\alias{.admm_lasso_cov_fast}
\alias{.admm_adalasso_fast}
\alias{.admm_adalasso_cov_fast}
\alias{.admm_ovglasso}
\alias{.admm_ovglasso_smo}
\alias{.admm_ovglasso_cov}
\alias{.admm_ovglasso_cov_smo}
\alias{.admm_ovglasso_fast}
\alias{.admm_ovglasso_smo_fast}
\alias{.admm_ovglasso_cov_fast}
\alias{.admm_ovglasso_cov_smo_fast}
\alias{.admm_spovglasso}
\alias{.admm_spovglasso_cov}
\alias{.admm_spovglasso_fast}
\alias{.admm_spovglasso_cov_fast}
\alias{.forward_diff_penalty_matrix}
\alias{.forward_diff_difference_matrix}
\title{Compute the discrete difference penalty matrix Q = D^T D}
\usage{
.lm_dof_LASSO_1lambda(
  X,
  coeff,
  lambda,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.lm_dof_OVGLASSO_1lambda(
  X,
  coeff,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.lm_cov_dof_OVGLASSO_1lambda(
  X,
  Z,
  coeff_X,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.lm_adaptive_dof_OVGLASSO_1lambda(
  X,
  coeff,
  coeff_LS,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.lm_cov_adaptive_dof_OVGLASSO_1lamba(
  X,
  Z,
  coeff_X,
  coeff_X_LS,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_dof_1lambda(
  W,
  coeff,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_cov_dof_1lambda(
  W,
  Z,
  coeff_W,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_smo_dof_1lambda(
  W,
  coeff,
  lambda,
  lambda2,
  diff_order,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_cov_smo_dof_1lambda(
  W,
  Z,
  coeff_W,
  lambda,
  lambda2,
  diff_order,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_adaptive_dof_1lambda(
  W,
  coeff,
  coeff_LS,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_cov_adaptive_dof_1lambda(
  W,
  Z,
  coeff_W,
  coeff_W_LS,
  lambda,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_smo_adaptive_dof_1lambda(
  W,
  coeff,
  coeff_LS,
  lambda,
  lambda2,
  diff_order,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_adaptive_cov_smo_dof_1lambda(
  W,
  Z,
  coeff_W,
  coeff_W_LS,
  lambda,
  lambda2,
  diff_order,
  GRmat,
  group_weights,
  var_weights,
  Uvec,
  err_primal,
  err_dual,
  rho,
  toler_c,
  toler_d
)

.f2s_bic(y, W, coeff_path, df)

.f2s_cov_bic(y, W, Z, W_coeff_path, Z_coeff_path, df)

.f2s_ebic(y, X, coeff_path, df, kappa)

.f2s_cov_ebic(y, W, Z, W_coeff_path, Z_coeff_path, df, kappa)

.admm_glasso(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  u,
  z,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_glasso_cov(
  W,
  Z,
  y,
  u,
  z,
  groups,
  group_weights,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_glasso_fast(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_glasso_cov_fast(
  W,
  Z,
  y,
  groups,
  group_weights,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spglasso(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  u,
  z,
  lambda1,
  lambda2,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spglasso_cov(
  W,
  Z,
  y,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  u,
  z,
  lambda1,
  lambda2,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spglasso_fast(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  lambda,
  alpha,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spglasso_cov_fast(
  W,
  Z,
  y,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  lambda,
  alpha,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_lasso(
  A,
  b,
  u,
  z,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_lasso_cov(
  W,
  Z,
  y,
  u,
  z,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_adalasso(
  A,
  b,
  var_weights,
  u,
  z,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_adalasso_cov(
  W,
  Z,
  y,
  var_weights,
  u,
  z,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_lasso_fast(
  A,
  b,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_lasso_cov_fast(
  W,
  Z,
  y,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_adalasso_fast(
  A,
  b,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_adalasso_cov_fast(
  W,
  Z,
  y,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  u,
  z,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_smo(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  u,
  z,
  lambda,
  lambda2,
  diff_order,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_cov(
  W,
  Z,
  y,
  u,
  z,
  groups,
  group_weights,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_cov_smo(
  W,
  Z,
  y,
  u,
  z,
  groups,
  group_weights,
  var_weights,
  lambda,
  lambda2,
  diff_order,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_fast(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_smo_fast(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  lambda,
  lambda2,
  diff_order,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_cov_fast(
  W,
  Z,
  y,
  groups,
  group_weights,
  var_weights,
  lambda,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_ovglasso_cov_smo_fast(
  W,
  Z,
  y,
  groups,
  group_weights,
  var_weights,
  lambda,
  lambda2,
  diff_order,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spovglasso(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  u,
  z,
  lambda,
  alpha,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spovglasso_cov(
  W,
  Z,
  y,
  u,
  z,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  lambda,
  alpha,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spovglasso_fast(
  A,
  b,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  lambda,
  alpha,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.admm_spovglasso_cov_fast(
  W,
  Z,
  y,
  groups,
  group_weights,
  var_weights,
  var_weights_L1,
  lambda,
  alpha,
  rho_adaptation,
  rho,
  tau,
  mu,
  reltol,
  abstol,
  maxiter,
  ping
)

.forward_diff_penalty_matrix(n, h, d)

.forward_diff_difference_matrix(n, h, d)
}
\arguments{
\item{n}{Length of the vector the operator acts on.}

\item{h}{Step size (default = 1.0).}

\item{d}{Derivative order (integer >= 1).}
}
\value{
Symmetric n x n matrix Q = D^T D.
}
\description{
Constructs the d-th order forward-difference operator D (scaled by 1/h^d)
and returns Q = D^T D.
}
\examples{
#Q1 <- forward_diff_penalty_matrix(5, h = 1, d = 1)
#Q2 <- forward_diff_penalty_matrix(5, h = 1, d = 2)
#Q1
}
\keyword{internal}
