% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_diff}
\alias{forward_diff}
\title{Forward finite difference approximation of order d}
\usage{
forward_diff(f, h, d)
}
\arguments{
\item{f}{Numeric vector containing function values.}

\item{h}{Step size (default = 1.0).}

\item{d}{Order of the derivative (integer >= 1).}
}
\value{
Numeric vector of forward differences of order d (length n, last d elements = NA).
#@examples
#f <- c(0, 1, 4, 9, 16)
#forward_diff(f = f, h = 1, d = 2)
}
\description{
Forward finite difference approximation of order d
}
