% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_auxiliary_fun.R
\name{softhresh}
\alias{softhresh}
\title{Function to solve the soft thresholding problem}
\usage{
softhresh(x, lambda)
}
\arguments{
\item{x}{the data value.}

\item{lambda}{the lambda value.}
}
\value{
the solution to the soft thresholding operator.
}
\description{
Function to solve the soft thresholding problem
}
\references{
\insertRef{hastie_etal.2015}{fdaSP}
}
